pyproject.toml
[project]
name = "Journal-app"
version = "0.1.0"
description = "Journal Backend"
authors = [
    {name = "Rivaldi Fadlan", email = "rivaldifadlan1@gmail.com"}
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "sqlmodel",
    "psycopg2-binary",
    "alembic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "argon2-cffi",
    "Pillow",
    "pytest",
    "httpx",
    "pyjwt",
    "python-multipart"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest", "httpx", "black", "isort"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["app"]
